+++ evaluation_scripts/run_example_evaluation.py
@@ -44,13 +44,13 @@ from absl import app
 from absl import flags
 import apache_beam as beam
 import numpy as np
-from weatherbenchX import aggregation
-from weatherbenchX import beam_pipeline
-from weatherbenchX import binning
-from weatherbenchX import time_chunks
-from weatherbenchX import weighting
-from weatherbenchX.data_loaders import xarray_loaders
-from weatherbenchX.metrics import deterministic
+from weatherbenchAfrica import aggregation
+from weatherbenchAfrica import beam_pipeline
+from weatherbenchAfrica import binning
+from weatherbenchAfrica import time_chunks
+from weatherbenchAfrica import weighting
+from weatherbenchAfrica.data_loaders import xarray_loaders
+from weatherbenchAfrica.metrics import deterministic
 
 
 _DEFAULT_VARIABLES = [
