+++ weatherbenchAfrica/binning_test.py
@@ -16,9 +16,9 @@ from importlib import resources
 from absl.testing import absltest
 from absl.testing import parameterized
 import numpy as np
-from weatherbenchX import binning
-from weatherbenchX import test_utils
-from weatherbenchX.data_loaders import sparse_parquet
+from weatherbenchAfrica import binning
+from weatherbenchAfrica import test_utils
+from weatherbenchAfrica.data_loaders import sparse_parquet
 import xarray as xr
