+++ weatherbenchAfrica/data_loaders/latency_wrappers_test.py
@@ -15,9 +15,9 @@
 
 from absl.testing import absltest
 import numpy as np
-from weatherbenchX import test_utils
-from weatherbenchX.data_loaders import latency_wrappers
-from weatherbenchX.data_loaders import xarray_loaders
+from weatherbenchAfrica import test_utils
+from weatherbenchAfrica.data_loaders import latency_wrappers
+from weatherbenchAfrica.data_loaders import xarray_loaders
 
 
 class LatencyWrappersTest(absltest.TestCase):
