+++ weatherbenchAfrica/metrics/wrappers.py
@@ -40,8 +40,8 @@ import abc
 from collections.abc import Sequence
 from typing import Any, Callable, Hashable, Iterable, Mapping, Union
 import numpy as np
-from weatherbenchX import xarray_tree
-from weatherbenchX.metrics import base
+from weatherbenchAfrica import xarray_tree
+from weatherbenchAfrica.metrics import base
 import xarray as xr
