+++ weatherbenchAfrica/test_data/metar-timeNominal-by-month/year=2020/month=1/2020-01.parquet
